refactor: move search tool logic to a new file
